Makefile Support
This patch adds support for a Makefile that makes building REX-Ray on a
clean system incredibly simple. Just type "make" and at the end of the
build a shiny new binary named "rexray" will be in $GOPATH/bin.

The best part is that if you unset the GOPATH variable in your shell,
the makefile will create a new, clean, isolated build environment for
you in the .build directory relative to the current directory. This
means all you  have to do to get a fresh build each time is remove the
.build  directory, and you can test building REX-Ray from scratch, in
isolation from any other Go environment on your system.