[32] Add commit hash to 'rexray version'
This patch changes the output of the 'rexray version' command to include
additional information about the REX-ray version and sources used to
build it. For example:

  [0]akutz@poppy:rexray$ rexray version
  SemVer: 0.2.0-rc2+2
  Branch: feature/32-version-info
  Commit: 86d6ced73b7968cdf292aac11ac6f46c86e01088
  Formed: Thu, 10 Sep 2015 18:50:47 CDT

Additionally, if the VERSION file exists at the root of the repository
and the file contains a valid semantic version, then that version's
major.minor.patch-preReleaseTag components will be used in lieu of those
grokked from a call to 'git describe --long' at build time.