[36] Broken Service
This patch fixes the issue #36 where changes to the infrastructure
inadvertently broke the service initialization files installed by
'rexray service install'.

Additionally this patch updates the 'rexray service install' command so
that the binary is no longer installed in `/opt/rexray'. Instead it is
up to the user to locate the binary in the desired location prior to
invoking 'rexray service install', at which point the service will be
installed, linking to the binary in its current location.