Release Tag Version Fix
This patch fixes an issue with the SemVers created from release tags as
those produced values should not include the build suffix of "+0" to
indicate there are zero commits since the last annotated tag.
Additionally all version calculation is now performed in the Makefile
and the "+dirty" suffix is appended to the SemVer and tarballs if the
code is built from a dirty repo.