Uname Uploads
Per issue #51, it would be useful to enable some way to access the most
recent build artifacts from a static path. This will be possible with
GitHub Releases with Release builds, but for the HEAD builds, this patch
enables a static resource to access artifacts built from the most recent
commit to master.