Curl Install Workflow Modification
This patch changes the behavior of the curl installation workflow.
While end-users will still execute the following:

    curl -sSL https://dl.bintray.com/emccode/rexray/install | sh -

The actual meat-and-potatoes of the installation routine is in the
file at `https://dl.bintray.com/emccode/rexray/install.sh`. In fact,
the above curl command simply contains the following content:

    #!/bin/sh

    # install rexray with the following command:
    #
    #   curl -sSL https://dl.bintray.com/emccode/rexray/install | sh -

    # get the real installer -- the current file is just an abstraction
    # to maintain a consistent download count for installs via the curl
    # install method
    curl -sSL https://dl.bintray.com/emccode/rexray/install.sh | sh -

Because updates to the installation script would requier deleting the
previous script, download counts would be eradicated with each update.
This level of abstraction enables a cumulative download count without
it ever having to be reset.