Repos, Installer, Pkg, REXRAY_HOME, & Svc Logging
This patch combines five changes:

1. The `stupid` repository has been renamed to `unstable`
2. A standard installation/uninstallation system has been created
   that is used regardless whether executed manually or by a package
   manager.
3. Support for building and deploying RPM and DEB packages
4. The `REX-Ray` binary may live wherever, but all of the data is
   created at the location defined by the environment variable
   `REXRAY_HOME`.
5. When executed as a service REX-Ray logs to
   $REXRAY_HOME/var/log/rexray.log

There's not much more to say but that the repository once known as
`stupid` has been renamed to `unstable`. The old `stupid` links will
continue to work for versions pushed to the repository when it was
named such, but once the next release candidate is pushed, those
versions will be purged.

The `REX-Ray` binary now supports the `install` and `uninstall`
commands, executed as `rexray install` or `rexray uninstall`. These
commands install the `REX-Ray` service if on a Linux system. Additionally,
the `install` and `uninstall` commands are also what is used by the RPM and
DEB packages, meaning that the `REX-Ray` installer is distribution and package
agnostic.

The Makefile now has support for the following targets:

* rpm
* rpm-all
* rpm-linux-amd64
* rpm-linux-386
* deb
* deb-all
* deb-linux-amd64

The above targets do what they imply they do -- they assemble the REX-Ray
binary into an RPM or DEB package. Please note the above targets expect
the appropriate architecture of REX-Ray to already be built and will not
force a new build of REX-Ray.

Also, it's not without *much* trouble that building a 32-bit DEB package is
possible on a 64-bit system, and thus the 32-bit DEB package is not avaialbe
at this time. Since `alien` was used to create the DEB packages from the RPMS,
the 32-bit RPM could easily be converted to a 32-bit DEB if a 32-bit Debian
system were available.

The packages install `REX-Ray` and register its service, but the server is
not started.

The `REX-Ray` binary can exist anywhere on a system, but when it is executed
it assumes the following paths (and will create them if they do not exist):

* `$REXRAY_HOME/etc/rexray`
* `$REXRAY_HOME/var/lib/rexray`
* `$REXRAY_HOME/var/log/rexray`
* `$REXRAY_HOME/var/run/rexray`

If the `REXRAY_HOME` environment variable does not exist then the above paths
will exist within the filesystem layout standard for both almost any UNIX-
derived system such as Linux or Darwin (OS X).

However, defining `REXRAY_HOME` to be `/opt/rexray/0.2.0` means that all four
paths above will be created and expected to exist at `/opt/rexray/0.2.0`.

Again, the `REX-Ray` binary can reside in `/usr/bin`,
`/opt/rexray/0.2.0/usr/bin`, at `/`, or anywhere on the filesystem. The binary's
location is independent of where the data is stored.

When `REX-Ray` is executed as a service the logs are no longer emitted to
`stdout` (except for `SystemD` where they are, but not limited to there).
Instead the logs are written to `$REXRAY_HOME/var/log/rexray.log`.

Of note, when logs are emitted to `stdout` they are color-coded, but not so when
written to a log file.