Spf13 Vendoring Update
This patch updates the vendoring of REX-Ray with regards to the Spf13
projects Viper and Cobra. The vendored Viper dependency was removed as
the pull request for the required feature was accepted into Viper
master. The vendored Cobra reference remains, but the Cobra fork was
rebased off of its master and necessary changes were made to REX-Ray to
accommodate the changes to Cobra.