The install.sh script checks the UID/Sudo req
This patch updates the install.sh script to only call /usr/bin/sudo if
the UID of the person executing the script is not 0 (root). This way if
the script is being executed as root or by sudo externally, sudo will
not be used again. This is very important on systems that don't support
TTYs or PTTYs.

Many thanks to @drumulonimbus for pointing this out.