Install Script Enhancements
This patch supplies multiple enhacnemnts to the curl-installs script,
supoorting the installation of the latest version version of REX-Ray
for CentOS, Ubuntu, CoreOS, and Darwin using RPMs, DEBs, and tarballs.

To install the latest version of REX-Ray simply execute:

  curl -sSL https://dl.bintray.com/emccode/rexray/install | sh -

However, this script will also allow users to install specific
versions of REX-Ray and even supports the discovery of those versions.
for example, to list the available REX-Ray packages use the "list"
command like so:

  curl -sSL https://dl.bintray.com/emccode/rexray/install | sh -s list -

This will emit the following curl commands to demonstrate how to
install one or more of the available packages, "unstable", "staged",
and "stable".

    curl -sSL https://dl.bintray.com/emccode/rexray/install | sh -s unstable -
    curl -sSL https://dl.bintray.com/emccode/rexray/install | sh -s staged -
    curl -sSL https://dl.bintray.com/emccode/rexray/install | sh -s stable -

Again, the "list" command outputs the curl commands to perform
installations, not how to traverse deeper into the package structure.
however, it's not hard to do that traversal. for example, to list the
contents of the "stable" package we'd execute:

  curl -sSL https://dl.bintray.com/emccode/rexray/install | sh -s list stable -

The above command would emit somthing similar to:

    curl -sSL https://dl.bintray.com/emccode/rexray/install | sh -s stable 0.2.0 -
    curl -sSL https://dl.bintray.com/emccode/rexray/install | sh -s stable latest -

As you can see, to traverse the structure, we simply insert the
command "list" as the first argument after the "-s" flag.