REX-Ray Core Testing Enhancements
This patch adds the root-level "test" package which exists solely to
provide coverage for the packages "github.com/emccode/rexray" and
"github.com/emccode/rexray/core". Eventually the "test" package will
also provide coverage for the CLI package as well.

Because of the way drivers are loaded, it's not possible for the core
package to share mock drivers with any other package. Thus in order
to prevent duplicate mock drivers, a single test package to provide
coverage of packages requiring mock drivers has been created.