CLI Refactored as Instance
This patch refactors the CLI so that multiple CLI instances can be
instantiated. This is to support both testing and a client/server
relationship. There are also package-level functions to directly
execute the CLI using a one-off instance. For example:

    import "github.com/emccode/rexray/rexray/cli"

    cli.ExecuteWithArgs("volume", "get", "-f", "json")

The above will get all the volumes and emit them in JSON format.

Although not yet integrated into code coverage, you can execute some
minimal CLI tests with the following command from the root of the
project:

    go test -run 'TestVolumeGetYaml|JSON' -v ./test/

It should look like this:

    [0]akutz@pax:rexray$ go test -run 'TestVolumeGetYaml|JSON' -v ./test/
    === RUN   TestVolumeGetYaml
    - name: test
      volumeid: test
      availabilityzone: test
      status: ""
      volumetype: ""
      iops: 0
      size: ""
      networkname: ""
      attachments: []

    --- PASS: TestVolumeGetYaml (0.00s)
    === RUN   TestVolumeGetJSON
    [{"Name":"test","VolumeID":"test","AvailabilityZone":"test","Status":"","VolumeType":"","IOPS":0,"Size":"","NetworkName":"","Attachments":null}]
    --- PASS: TestVolumeGetJSON (0.00s)
    PASS
    ok      github.com/emccode/rexray/test  0.016s