Gofig Integration / Namespace Change
This patch does away with the core/config package and instead uses the
Gofig project for all of the configuration responsibilities.
Additionally all of the REX-Ray properties are now under the rexray
namespace in the YAML configuration. What was once "host" is now
"rexray.host." These changes are reflected in the documentation.