Fix for not setting Gofig config dirs (Rebased)
This patch was originally commit
d04023572e47e5c18fa549fe7f7de301e8470c63, but it failed to build in CI
as it was not rebased off of master prior to being merged. This is the
rebased patch.

This patch fixes the issue where the Gofig's global (/etc) and user
($HOME) directories were not set, resulting in the default config files
not being loaded.