Bugfix for change in Glide behavior
This patch fixes an issue where Glide no longer seems to handle
transitive dependencies correctly unless the referencing dependency is
ordered above another that may have a conflict.

In this instance the dependency upon `github.com/akutz/goof` was moved
to the top of the `glide.yaml` file so that its dependency upon a fork
of the `logrus` project would be honored when glide evaluated the
transitive dependency chain.