Added support for Volume Driver with Docker 1.10
This commit introduces new functionality that is exposed to the
Docker Volume Driver in Engine 1.10.  Calls for Get/List are
now implemented to enable the engine to rely on volume driver
to return list of existing volumes.  A list locally is no longer
stored which means the array is the single source of truth
for available volumes.