Config-Based Module Support
This patch adds support for configuration modules via the REX-Ray
configuration source. For example, the following configuration is loaded
by default:

    rexray:
        modules:
            default-admin:
                type:     admin
                desc:     The default admin module.
                host:     tcp://127.0.0.1:7979
                disabled: false
            default-docker:
                type:     docker
                desc:     The default docker module.
                host:     unix:///run/docker/plugins/rexray.sock
                spec:     /etc/docker/plugins/rexray.spec
                disabled: false

The above configuration indicates that there are two modules that need
to be initialized and started:

  1. default-admin
  2. default-docker

Default modules can be overridden in a custom configuration file by
simply using the same names as above.

If the `host` or `spec` properties are not defined for a docker module,
then the sanitized name of the module is used to build the paths to a
socket and spec file. For example:

    rexray:
        modules:
            "isilon 2":
                type: docker

The above configuration would create a Docker module hosted using a
socket file at `unix:///run/docker/plugins/isilon-2.sock` and spec file
at `/etc/docker/plugins/isilon-2.spec`.

If the configuration file is used to override part of the
`default-docker` module's configuration, for example its description,
it's also possible to omit the `default-docker` module's `host` and
`spec` properties as well. The difference is that the `default-docker`
module's `host` and `spec` properties will default to values based not
on the module name but `unix:///run/docker/plugins/rexray` and
`/etc/docker/plugins/rexray.spec`.

Modules can also be disabled using `disabled: true` with any module,
including default modules. This is useful as it allows the disabling of
a module without completely removing its configuration from the file.