Docs Evolved
This patch updates the documentation to be more concise and succinct -
a reorganization effort for both aesthetic and content reasons. The
top-level Automation page formerly under User Guide is now a sub-section
of the Installation page. The Storage Providers menu is no longer full
of sub-menus, but a single page.