The executor CLI
This patch introduces the executor CLI that is built for each driver. To
build executors use the following Make targets:

  - build-executor

      builds the executor for the system GOOS_GOARCH with the default
      driver (mock)

  - build-executors

      builds the executors for all drivers for the system GOOS_GOARCH

  - build-executors-all

      builds the executors for all drivers for all GOOS_GOARCH
      combinations

  - build-executor-GOOS_GOARCH

      builds the executor for the system GOOS_GOARCH with the default
      driver (mock)

To build an executor for a different driver or a different driver and
GOOS_GOARCH combination without using the mulit-build targets, invoke
the `build-executor` or `build-executor-GOOS_GOARCH` targets specifying
different go tags. For example:

  env GO_TAGS='ec2 driver' make build-executor

The above command will build the executor for the ec2 driver for the
system GOOS_GOARCH.