Introducing Asynchronous Tasks
This patch introduces the concept of Tasks to libStorage. For any
volumes or snapshots collection URL, if the query string `?async` is present,
libStorage will execute the intended operation in a goroutine, and immediately
return the HTTP status 202 (Accepted) to the client.

The 202 response will have a body that adheres to the JSON task schema:

    "task": {
        "type": "object",
        "properties": {
            "id": {
                "type": "number",
                "description": "The task's unique identifier."
            },
            "name": {
                "type": "string",
                "description": "The name of the task."
            },
            "user": {
                "type": "string",
                "description": "The name of the user that created the task."
            },
            "completeTime": {
                "type": "number",
                "description": "The time stamp (epoch) when the task was completed."
            },
            "queueTime": {
                "type": "number",
                "description": "The time stamp (epoch) when the task was created."
            },
            "startTime": {
                "type": "number",
                "description": "The time stamp (epoch) when the task started running."
            },
            "result": {
                "type": "object",
                "description": "The result of the operation."
            },
            "error": {
                "type": "object",
                "description": "If the operation returned an error, this is it."
            },
            "fields": { "$ref": "#/definitions/fields" }
        },
        "required": [ "id", "name",  "user", "queueTime" ],
        "additionalProperties": false
    }

The task's ID can be used to retrieve the result of an asynchronous operation.
To get the result of an operation, use the URL `/tasks/{taskID}`.

If an operation is not asynchronous but times out due to HTTP timeout rules,
the libStorage server will continue to process the operation in the background.
A 408 status code is returned along with a response body that is a JSON task.
The task's ID can be used to retrieve the result of the operation using the
aforementioned task URL.