Refactored Client Interfaces & Enhanced Tests
This patch refactors the Client interfaces such that ./api/client now
has the Client interface which is the base for the ./client/Client and
./api/client/APIClient interfaces. The ./api/client/APIClient interface
is what the ./api/client/Client interface used to be. The
./client/Client interface is a superset of the ./api/client/Client
functions with the addition of helper functions. However, the
./client/Client object returned by the NewClient function can be cast to
./api/client/APIClient when it's necessary to access underlying,
advanced API functions.

Finally the test harness has been enhanced to provided canned test
routines for testing executors, getting the instance ID, and validating
common API routes.