InstanceID and LocalDevice headers
This patch updates the libStorage client so that all outgoing HTTP
requests to the storage-related API functions automatically have added
to their payloads, as headers, the client's instance ID and local device
information.

The instance ID and local device data is read on the server-side and
made available to drivers with the context.