Added server run and synchronous run capabilities to server package
This commit moves CLI based server start methods to the server package.
This ensure that a standard approach to starting the services is possible.
The server can be started in a synch or async mode and allows configuration
objects to be specified or a parameters to be passed that build a configuration.