Transactions
This patch adds loose-transactions, the ability to track the provenance
of an operation from the time it was generated on a client, through the
server round-trip, and back into the client.

It does this via two HTTP headers, libStorage-txID and libStorage-txCR.
The former is a UUID associated with every client->server->client
round-trip and the latter is the UTC epoch that is the time at which the
transaction was created.