Go-bindata Proper
This patch automatically fetches go-bindata as part of `glide.yaml` and
will build it from that source if it does not exist. Projects that
vendor libStorage should not have to do anything special with regards to
`go-bindata` as long as said projects are using Glide. The libStorage
Makefile's `executors` target which projects that embed libStorage
already execute now depends on the target that builds `go-bindata`.


Former-commit-id: b6641ce69ca8f1efcacf1fd939a81deb3fd6f9a9