Added timeout setting for module startup
This commit introduces a new startup timeout setting for modules.
In the case of pathcache on startup it is a heavy lifting
operation which requires the lookup of backend storage and mount
points. Depending on the driver, this may be a lenghty operation.
The timeout is now configurable and starts at 10 seconds instead
of 3.