Version Info & Time in Logs
This patch adds a new resource to the libStorage root resources,
/version. The /version resource returns the version information for the
libStorage server. For example:

  {
    "semver": "0.1.0+5+dirty",
    "shaLong": "9a4c28adb385281f1b053032e6fefb77f597b7cf",
    "buildTimestamp": 1463666979,
    "branch": "feature/version",
    "arch": "Darwin-x86_64"
  }

In addition, the version information is also available to the reference
server binary, lss-$GOOS. For example:

  $ lss-darwin --version
  Binary: /Users/akutz/Projects/go/bin/lss-darwin
  SemVer: 0.1.0+5+dirty
  OsArch: Darwin-x86_64
  Branch: feature/version
  Commit: 9a4c28adb385281f1b053032e6fefb77f597b7cf
  Formed: Thu, 19 May 2016 09:09:39 CDT

Finally, this patch also adds the field `time` to all context-emitted
log messages so that every entry includes `time=EPOCH` to indicate at
what time the log entry was created.