Help Feature
This patch adds the `/help` resource (replacing the `/version) resource.
The help resource has the child resources:

  /help/version
  /help/env
  /help/config

The /help/version resource is just the /version resource, relocated.

Both the /help/env and /help/config resources require the query string
key `admin` to be present in the resource URL with a value set to the
admin token that is emitted to the server-side log upon start up. If the
admin token is omitted from the resource URL then a 401 unauthorized
status is returned.

If the admin token is provided the /help/config resource prints the
runtime configuration in use by the server, formatted as JSON. The
/help/env resource prints the server process's environment variable
collection, also formatted as a JSON string.