Bugfix for Executor Stdout Logging
This patch fixes an issue where the XCLI harness responsible for calling
the executor binary was capturing combined Stdout & Stderr streams. As
it turns out, Logrus emits all log messages on the Stderr stream, so if
logging was verbose, it interrupted the logic for capturing the executor
result. Now the XCLI harness only looks at the Stdout stream.