Refactor Auto Bits
This patch removes all of the easy buttons from the client and server.
These functions were never part of the core design plan, and the use
cases the aforementioned functions fulfilled are still largely met due
to the planned improvements implemented in the libStorage core. The
removed bits had too many side effects, and were not congruent with the
project's overall design.