Fix for Volume.MountPoint()
This patch resolves issue #174 where the Volume.MountPoint() function
returns an empty string after the IntegrationDriver.Mount() function.
The real reason is because a second Volume inspection operation should
be executed to populate the mapping data, but to avoid the round-trip,
the mount point is now directly set on th volume's attachment object.