Remove CLI default-docker scope
This patch removes the root CLI scope of `default-docker`. Once upon a
time it was there to enable preemption and other volume related options
for CLI commands, but those options are no longer configured there.

The presence of the `default-docker` scope as a root CLI scope was
causing issues due to the libStorage host and other special properties
being misinterpreted when defined in the `default-docker` section, even
when the `default-docker` module was disabled.