Task Service Memory Fix
This patch addresses an issue with the task service keeping the results
of operations with no expiration. Because the results are kept in
memory, and because the internal implementation of a task includes
references to the HTTP request object, the context, etc., hardly any
memory is ever freed from a single API call. This results in the
libStorage server consuming upwards of 50GB of RAM after several hours
of constant usage.

This patch includes two, new configuration keys:

    libstorage:
      server:
        tasks:
          exeTimeout: 60s
          logTimeout: 0

The key `libstorage.server.tasks.exeTimeout` defaults to one minute, and
it is the duration of time in which as task is expected to complete
before being timed out and backgrounded, informing the user of the task
ID that can be used to track the operation.

The second key, `libstorage.server.tasks.logTimeout` defaults to zero
and is the duration of time for which a task is kept in memory after a
task completes successfully or otherwise. This value should be increased
to some positive duration if tasks often timeout and results need to be
obtained after the fact on a regular basis.

Both property values are expected to be parsed by the time.ParseDuration
(https://golang.org/pkg/time/#ParseDuration) function. A duration can be
60s, 1m, 60000ms, or even 1h.