Adds a No-SSL-Warning Installation Option
This patch adds a No-SSL-Warning installation option to the curl-bash
install script. Specifying the flag `INSECURE=1` ahead of the `sh`
command after the pipe in the `curl`-bash installation command now
causes the script's internal `curl` commands to specify the `-k` flag,
indicating that SSL warnings should be ignored.

For example, to list the available versions to be installed using the
new No-SSL-Warning option, the following command would be used:

        $ curl -ksSL https://dl.bintray.com/emccode/rexray/install | \
          INSECURE=1 sh -s -- list stable

Please note in the above example that the `-k` flag must be added
manually to the initial `curl` command on the CLI. Setting `INSECURE=1`
only instructs the `curl` commands *inside* the fetched install script
to use the `-k` flag, not the initial one.