Fix for Profiler Race Condition
The profiler added to REX-Ray worked fine until it was relocated to the
`init()` function from the `main()` function. This was done in hopes of
capturing the effects of loading the other packages. However, it ended
up stalling the application randomly, likely because of a race condition
created trying to load profiling when the profiling packages were not
yet loaded themselves.

This patch moves the profiler code back into the `main()` function.