Refine rexray volume ls filter
This commit fixes the following issues:
1. When rexray volume ls was used specifying only a volume ID,
either a volume with a blank name, or a volume matching the same
volume ID would be returned.
2. When rexray volume ls was used specifying a volume name that
matched multiple volumes, only one volume with a matching volume
name would be returned. In REX-Ray 0.3.0, all volumes that matched
the volume name would be returned.
3. String comparison is not optimized. This has now been changed
to use strings.EqualFold for comparing strings instead to avoid
unnecessary memory allocations with strings.ToLower as per
@benbjohnson's article Go Walkthrough: bytes + strings packages.