Fix to the VFS Test Issue
This patch should hopefully fix the VFS test issue where the underlying
AUFS performance causes the TestVolumeSnapshot VFS test to fail. This
change allows the point of failure to execute up to 10 times with a 1
second sleep period between each iteration. If at the end of 10
iterations it hasn't succeeded, the test will fail.