NextDevice Fix
This patch fixes an issue where the NextDevice function of the Executor
was not being invoked prior to calling VolumeAttach. No value was being
provided in the VolumeAttachOpts.NextDevice field. Now the executor
should properly gather that information and provide it to the server in
the event of an attach call.